Working on modular monolith
