+++ customers/src/main/java/customers/CustomerController.java
@@ -1,4 +1,4 @@
-package web;
+package customers;
 
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import jakarta.validation.Valid;
