+++ persistence/src/main/java/persistance/repository/ExpenseRepository.java
@@ -1,12 +1,14 @@
-package web;
+package persistance.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.query.Param;
+import org.springframework.stereotype.Repository;
 import persistance.entity.Expense;
 
 import java.util.List;
 import java.util.Optional;
 
+@Repository
 public interface ExpenseRepository extends JpaRepository<Expense, Long> {
     Optional<List<Expense>> findAllByVehicleId(@Param("id") Long id);
 }
