+++ starter/src/main/java/com/fueltracker/FuelTrackerApplication.java
@@ -2,9 +2,8 @@ package com.fueltracker;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import web.CustomerService;
 
-@SpringBootApplication(scanBasePackageClasses = {CustomerService.class}, scanBasePackages = "com.fuelTracker")
+@SpringBootApplication(scanBasePackages = {"com.fuelTracker", "customers", "vehicles", "refuel", "expense", "totalcost"})
 public class FuelTrackerApplication {
     public static void main(String[] args) {
         SpringApplication.run(FuelTrackerApplication.class, args);
