+++ totalCost/src/main/java/totalcost/utils/DistanceCalculator.java
@@ -1,4 +1,4 @@
-package com.fueltracker.utils;
+package totalcost.utils;
 
 
 import persistance.entity.Expense;
